temp commit for floating app values

+++ fdp-client/src/main/java/fk/retail/ip/fdp/model/FdpRequirementEntityData.java
@@ -21,7 +21,7 @@ public class FdpRequirementEntityData {
     String forecast;
     double quantity;
     String supplier;
-    Integer app;
+    double app;
     Integer mrp;
     String currency;
     Date requiredByDate;
