+++ requirement/src/main/java/fk/retail/ip/requirement/model/RequirementDownloadLineItem.java
@@ -58,7 +58,7 @@ public class RequirementDownloadLineItem {
     @JsonProperty("Forecast")
     private String forecast;
     @JsonProperty("Total Value")
-    private Integer totalValue;
+    private Double totalValue;
     @JsonProperty("Intransit")
     private Integer intransitQty;
     @JsonProperty("Quantity")
@@ -68,7 +68,7 @@ public class RequirementDownloadLineItem {
     @JsonProperty("MRP")
     private Integer mrp;
     @JsonProperty("Purchase Price")
-    private Integer app;
+    private Double app;
     @JsonProperty("Currency")
     private String currency;
     @JsonProperty("SLA")
@@ -100,7 +100,7 @@ public class RequirementDownloadLineItem {
     @JsonProperty ("CDO Quantity Override Reason")
     private String cdoQuantityOverrideReason;
     @JsonProperty ("CDO Price Override")
-    private Integer cdoPriceOverride;
+    private Double cdoPriceOverride;
     @JsonProperty ("CDO Price Override Reason")
     private String cdoPriceOverrideReason;
     @JsonProperty ("New SLA")
